[Agent]
Refactor Agent: SafeHandle token cleanup, CancellationToken usage, and uint process IDs

- Changed Agent.ImpersonationToken from IntPtr to SafeTokenHandle to ensure
  automatic handle disposal on reassignment.
- Updated StealTokenCommand, MakeTokenCommand, RevertToSelfCommand and all
  token-consuming commands to use SafeTokenHandle.
- Changed ProcessCreationParameters.Token to SafeTokenHandle; CreateProcess
  wrappers now call DangerousGetHandle() at the native API boundary.
- Added CancellationToken support to APIWrapper.ReadPipeToEnd; removed
  Thread.Abort() from ShellCommand and ForkAndRunCommand.
- Replaced Thread.Sleep with await Task.Delay in PsExecCommand.
- Wrapped SCManager/service handles in SafeServiceHandle in PsExecCommand.
- Removed obsolete int overloads in APIWrapper; process IDs are now uniformly
  uint across StealToken, OpenProcess and ParentProcessId.

+++ Agent/Commands/Execution/InjectCommand.cs
@@ -35,7 +35,7 @@ namespace Agent.Commands
             //    return;
             //}
 
-            IntPtr hProcess = APIWrapper.OpenProcess(processId, ProcessAccessFlags.PROCESS_VM_WRITE |
+            IntPtr hProcess = APIWrapper.OpenProcess((uint)processId, ProcessAccessFlags.PROCESS_VM_WRITE |
     ProcessAccessFlags.PROCESS_VM_OPERATION |
     ProcessAccessFlags.PROCESS_CREATE_THREAD);
             // OUVRIR UN NOUVEAU HANDLE AVEC LES BONS DROITS
